requirements.txt
matplotlib
cython
scipy
lalsuite
numpy
torch
appdirs
h5py
torch-summary
regex
